+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/docx_paragraph/docx_paragraph_run/docx_paragraph_run_helpers.rb
@@ -33,7 +33,7 @@ module OoxmlParser
         when 'effect'
           self.effect = node_child.attribute('val').value
         when 'position'
-          self.position = (node_child.attribute('val').value.to_f / (28.0 + 1.0 / 3.0) / 2.0).round(1)
+          self.position = (node_child.attribute('val').value.to_f / (28.0 + (1.0 / 3.0)) / 2.0).round(1)
         when 'em'
           self.em = node_child.attribute('val').value
         when 'spacing'